Sync Fork from Upstream Repo

+++ src/install/installer.ts
@@ -17,11 +17,10 @@
 import fs from 'fs';
 import path from 'path';
 import util from 'util';
-import removeFolder from 'rimraf';
 import lockfile from 'proper-lockfile';
 import {Registry, allBrowserNames, isBrowserDirectory, BrowserName, registryDirectory} from '../utils/registry';
 import * as browserFetcher from './browserFetcher';
-import { getAsBooleanFromENV, calculateSha1 } from '../utils/utils';
+import { getAsBooleanFromENV, calculateSha1, removeFolders } from '../utils/utils';
 
 const fsMkdirAsync = util.promisify(fs.mkdir.bind(fs));
 const fsReaddirAsync = util.promisify(fs.readdir.bind(fs));
@@ -29,11 +28,10 @@ const fsReadFileAsync = util.promisify(fs.readFile.bind(fs));
 const fsExistsAsync = (filePath: string) => fsReadFileAsync(filePath).then(() => true).catch(e => false);
 const fsUnlinkAsync = util.promisify(fs.unlink.bind(fs));
 const fsWriteFileAsync = util.promisify(fs.writeFile.bind(fs));
-const removeFolderAsync = util.promisify(removeFolder);
 
 const PACKAGE_PATH = path.join(__dirname, '..', '..');
 
-export async function installBrowsersWithProgressBar(browserNames: BrowserName[] = allBrowserNames) {
+export async function installBrowsersWithProgressBar(browserNames: BrowserName[] = Registry.currentPackageRegistry().installByDefault()) {
   // PLAYWRIGHT_SKIP_BROWSER_DOWNLOAD should have a value of 0 or 1
   if (getAsBooleanFromENV('PLAYWRIGHT_SKIP_BROWSER_DOWNLOAD')) {
     browserFetcher.logPolitely('Skipping browsers download because `PLAYWRIGHT_SKIP_BROWSER_DOWNLOAD` env variable is set');
@@ -76,7 +74,7 @@ async function validateCache(linksDir: string, browserNames: BrowserName[]) {
       linkTarget = (await fsReadFileAsync(linkPath)).toString();
       const linkRegistry = new Registry(linkTarget);
       for (const browserName of allBrowserNames) {
-        if (!linkRegistry.shouldDownload(browserName))
+        if (!linkRegistry.shouldRetain(browserName))
           continue;
         const usedBrowserPath = linkRegistry.browserDirectory(browserName);
         const browserRevision = linkRegistry.revision(browserName);
@@ -95,21 +93,20 @@ async function validateCache(linksDir: string, browserNames: BrowserName[]) {
   }
 
   // 2. Delete all unused browsers.
-  let downloadedBrowsers = (await fsReaddirAsync(registryDirectory)).map(file => path.join(registryDirectory, file));
-  downloadedBrowsers = downloadedBrowsers.filter(file => isBrowserDirectory(file));
-  const directories = new Set<string>(downloadedBrowsers);
-  for (const browserDirectory of usedBrowserPaths)
-    directories.delete(browserDirectory);
-  for (const directory of directories) {
-    browserFetcher.logPolitely('Removing unused browser at ' + directory);
-    await removeFolderAsync(directory).catch(e => {});
+  if (!getAsBooleanFromENV('PLAYWRIGHT_SKIP_BROWSER_GC')) {
+    let downloadedBrowsers = (await fsReaddirAsync(registryDirectory)).map(file => path.join(registryDirectory, file));
+    downloadedBrowsers = downloadedBrowsers.filter(file => isBrowserDirectory(file));
+    const directories = new Set<string>(downloadedBrowsers);
+    for (const browserDirectory of usedBrowserPaths)
+      directories.delete(browserDirectory);
+    for (const directory of directories)
+      browserFetcher.logPolitely('Removing unused browser at ' + directory);
+    await removeFolders([...directories]);
   }
 
   // 3. Install missing browsers for this package.
-  const myRegistry = new Registry(PACKAGE_PATH);
+  const myRegistry = Registry.currentPackageRegistry();
   for (const browserName of browserNames) {
-    if (!myRegistry.shouldDownload(browserName))
-      continue;
     await browserFetcher.downloadBrowserWithProgressBar(myRegistry, browserName).catch(e => {
       throw new Error(`Failed to download ${browserName}, caused by\n${e.stack}`);
     });
